+++ tests/unit/aggregation/test_nash_mtl.py
@@ -1,7 +1,8 @@
-import pytest
+from utils.optional_deps import skip_if_deps_not_installed
 
-pytest.importorskip("cvxpy")
-pytest.importorskip("ecos")
+from torchjd.aggregation._nash_mtl import _NashMTLWeighting
+
+skip_if_deps_not_installed(_NashMTLWeighting)
 
 from pytest import mark, raises
 from torch import Tensor
@@ -9,7 +10,6 @@ from torch.testing import assert_close
 from utils.tensors import ones_, randn_, tensor_
 
 from torchjd.aggregation import NashMTL
-from torchjd.aggregation._nash_mtl import _NashMTLWeighting
 
 from ._asserts import assert_expected_structure, assert_non_differentiable
 from ._inputs import nash_mtl_matrices