+++ CHANGELOG.md
@@ -8,6 +8,14 @@ changelog does not include internal changes that do not affect the user.
 
 ## [Unreleased]
 
+### Changed
+
+- **BREAKING**: Removed `numpy`, `quadprog` and `qpsolvers` from the main dependencies of `torchjd`,
+  (which now only has `torch` as its main dependency). This makes the base version of `torchjd`
+  (installed with `pip install torchjd`) much lighter, but it means that users of `UPGrad` and
+  `DualProj` now have to install the new optional dependency group `quadprog_projector` explicitly
+  (with e.g. `pip install "torchjd[quadprog_projector]"`).
+
 ## [0.11.0] - 2026-05-18
 
 ### Changed